requirements.txt
flask==2.0.3
werkzeug==2.0.3
python-dotenv==0.19.2
requests==2.27.1
translate==3.6.1